refactor: separate InstallButtonComponent
